CHATME-66 Made sidebar layout
